working on particles not being deleted properly

+++ source/SOAPoint.h
@@ -1,14 +1,14 @@
 #ifndef __SOA_POINT_H
 #define __SOA_POINT_H
 
-#include <cstdlib>
+//#include <cstdlib>
 
 typedef struct
 {
   float *x;
   float *y;
   float *z;
-  size_t size;
+  unsigned int size;
 } soa_point_t;
 
 #endif
